in and not in always return bool
